split parse.c
