+++ src/codegen.c
@@ -60,15 +60,15 @@ void gen(Node *node){
 			printf("	push rdi\n");
 			return;
 		case ND_IF:
-			// adjust rsp
 			printf("	push rax\n");
 			gen(node->lhs);
+
 			printf("	pop rax\n");
 			printf("	cmp rax,0\n");
 			printf("	je .Lend%03d\n",label_end);
-			// adjust rsp
 			printf("	pop rax\n");
 			gen(node->rhs);
+
 			printf(".Lend%03d:\n",label_end);
 			label_end++;
 			return;
@@ -78,10 +78,12 @@ void gen(Node *node){
 			printf("	pop rax\n");
 			printf("	cmp rax,0\n");
 			printf("	je .Lelse%03d\n",label_else);
+
 			// expr in if
 			gen(node->rhs->lhs);
 			printf("	jmp .Lend%03d\n",label_end);
 			printf(".Lelse%03d:\n",label_else);
+
 			// expr in else
 			gen(node->rhs->rhs);
 			printf(".Lend%03d:\n",label_end);
