+++ src/main.c
@@ -2,7 +2,6 @@
 
 int main(int argc,char **argv){
 	int i,j;
-
 	char reg[6][4]={"rdi","rsi","rdx","rcx","r8","r9"};
 
 	if(argc!=2){
